Extract Integrity Checker
